requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.2
matplotlib==3.8.4
scikit-learn==1.4.2
authlib==1.3.0
uvicorn==0.29.0  # opcional
